DE-3erEntregable/requirements.txt
apache-airflow==2.3.3
pandas==2.0.3
sqlalchemy==1.4.49